Pagination Styling & Documentation
I began to style the pagination, but its a little over my head at the
moment.

I created documentation for TipJar as well as a page entitled TipJargon
which can list all the different terms related to applications like
Photoshop. I.E. describing what the "tool bar" is and where it is
located in Photoshop.